create objects outside of sandbox to avoid web app class loader deny
elsewhere OgnlSecurityManager denies access reading .class files when web app class loader tries to lazy load classes e.g. OgnlSandbox$1.class for inner anonymous class PrivilegedExceptionAction.
